storage: extract adjustable_semaphore
This was previously storage::adjustable_allowance, used
in storage_resources for concurrency limits that could
be adjusted at runtime via configuration bindings.

Renaming for clarity (as the noun 'allowance' didn't
really say anything that 'semaphore' doesn't say more
clearly).

This will also be used in cloud_storage for similar
purposese.
